+++ components/teamSelect.tsx
@@ -7,7 +7,7 @@ const { Option, OptGroup } = Select;
 interface TeamSelectProps {
 	teamsData: TeamSelectData[];
 	currentTeamID: string | undefined;
-	handleChange: Dispatch<SetStateAction<string | undefined>>;
+	handleChange: Dispatch<SetStateAction<string>>;
 }
 
 function withCheckMark(value: string) {
