+++ next.config.js
@@ -1,12 +1,3 @@
 module.exports = {
-	reactStrictMode: true,
-	async redirects() {
-		return [
-			{
-				source: '/',
-				destination: '/dashboard',
-				permanent: true,
-			},
-		];
-	},
+	reactStrictMode: true
 };
