+++ pages/api/team-management.ts
@@ -3,11 +3,11 @@ import { customAlphabet } from 'nanoid';
 import Team from '../../models/team';
 import dbConnect from '../../middleware/database';
 import { getSession } from 'next-auth/react';
-import { TeamProfile } from '../team';
 import User from '../../models/user';
 import { ObjectId } from 'mongodb';
 import { MongoServerError } from 'mongoose/node_modules/mongodb';
 import log from '../../middleware/log';
+import { TeamProfile } from '../../types/client';
 
 const nanoid = customAlphabet('abcdefghijklmnopqrstuvwxyz', 7);
