+++ pages/index.tsx
@@ -1,29 +1,31 @@
+import { Skeleton } from 'antd';
+import { Content } from 'antd/lib/layout/layout';
 import { signIn, signOut, useSession } from 'next-auth/react';
-import Outline from '../components/outline';
+import HackerDash from './HackerDash';
+import JudgeDash from './JudgeDash';
+import OrganizerDash from './OrganizerDash';
 
 export default function Page() {
 	const { data: session, status } = useSession();
 
 	return (
-		<Outline selectedKey="">
-			{!session && (
+		<Content style={{ padding: '30px' }}>
+			{ !session && status === "unauthenticated" && (
 				<>
 					Not signed in <br />
 					<button onClick={() => signIn()}>Sign in</button>
 				</>
-			)}
-			{session && (
+			) }
+			{ !session && status === "loading" && <Skeleton /> }
+			{ session && (
 				<>
 					Signed in as {session.user?.email} <br />
-					<button onClick={() => signOut()}>Sign out</button>
+					<button onClick={() => signOut()}>Sign out</button> <br /><br />
+					{ session.userType === "HACKER" && <HackerDash /> }
+					{ session.userType === "JUDGE" && <JudgeDash /> }
+					{ session.userType === "ORGANIZER" && <OrganizerDash />  }
 				</>
-			)}
-		</Outline>
+			) }
+		</Content>
 	);
-}
-
-export async function getStaticProps() {
-	return {
-		props: { title: 'Sign In' },
-	};
 }
